Add asciidoc.conf
Borrow an asciidoc.conf from something I did a number of years ago to
get the /BR/ replacement to insert a newline in the presentation
title.  This keeps it from wrapping on my monitor.

Modify the asciidoc command in the Makefile to use the new
configuration file.

Modify the main slideshow to use the replacement.